move FunctionDefinitionFormat to enums
