+++ aci/types/enums.py
@@ -16,3 +16,12 @@ class Visibility(str, Enum):
 
     PUBLIC = "public"
     PRIVATE = "private"
+
+
+class FunctionDefinitionFormat(str, Enum):
+    BASIC = "basic"  # name and description only
+    OPENAI = "openai"  # openai function call format (for the chat completions api)
+    OPENAI_RESPONSES = (
+        "openai_responses"  # openai function call format (for the responses api, the newest API)
+    )
+    ANTHROPIC = "anthropic"  # anthropic function call format
