+++ aci/types/functions.py
@@ -1,16 +1,8 @@
-from enum import Enum
 from typing import Any
 
 from pydantic import BaseModel
 
-
-class FunctionDefinitionFormat(str, Enum):
-    BASIC = "basic"  # name and description only
-    OPENAI = "openai"  # openai function call format (for the chat completions api)
-    OPENAI_RESPONSES = (
-        "openai_responses"  # openai function call format (for the responses api, the newest API)
-    )
-    ANTHROPIC = "anthropic"  # anthropic function call format
+from aci.types.enums import FunctionDefinitionFormat
 
 
 class GetFunctionDefinitionParams(BaseModel):
